Adding redux basic structure
