+++ webapp/package.json
@@ -13,9 +13,12 @@
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-flexbox-grid": "^1.0.2",
+    "react-redux": "^5.0.4",
     "react-router": "^3.0.0",
+    "react-router-redux": "^4.0.8",
     "react-scripts": "^0.9.5",
     "react-tap-event-plugin": "^2.0.1",
+    "redux": "^3.6.0",
     "style-loader": "^0.16.1",
     "webpack": "^2.4.1",
     "webpack-dev-server": "^2.4.2"
