+++ webapp/src/app/core/actions/index.js
@@ -1,0 +1,3 @@
+const getApiUrl = () => process.env.REACT_APP_API_URL || "http://localhost:3000";
+
+export default getApiUrl;
