+++ webapp/src/app/core/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import { routerReducer } from "react-router-redux"
+
+const beagleboxapp = combineReducers({
+  routing: routerReducer,
+})
+
+export default beagleboxapp
