+++ webapp/src/app/core/store.js
@@ -1,0 +1,5 @@
+import { createStore } from "redux"
+
+import reducers from "./reducers"
+
+export default createStore(reducers);
