+++ webapp/src/app/core/types/index.js
@@ -1,0 +1,9 @@
+function combineTypes(typeArray) {
+  return Object.assign({}, ...typeArray);
+}
+
+const types = combineTypes([
+
+]);
+
+export default types;
