+++ webapp/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { IntlProvider } from 'react-intl'
+import { Provider } from 'react-redux'
 import i18nData from './app/config/i18n.config'
 import { Router, Route, hashHistory, Redirect } from 'react-router'
 
@@ -11,20 +12,24 @@ import RequestLocation from './app/view/request'
 import Receive from './app/view/receive'
 import History from './app/view/history'
 
+import store from  "./app/core/store"
+
 import './index.css'
 
 ReactDOM.render(
-  <IntlProvider {...i18nData}>
-    <Router history={hashHistory}>
-      <Redirect from='/' to='/inicio' />
+  <Provider store={store}>
+    <IntlProvider {...i18nData}>
+      <Router history={hashHistory}>
+        <Redirect from='/' to='/inicio' />
 
-      <Route path='/' component={App}>
-        <Route path='inicio' component={Homepage}></Route>
-        <Route path='informacoes-gerais' component={GeneralInformation}></Route>
-        <Route path='solicitar' component={RequestLocation}></Route>
-        <Route path='receber' component={Receive}></Route>
-        <Route path='historico' component={History}></Route>
-      </Route>
-    </Router>
-  </IntlProvider>,
+        <Route path='/' component={App}>
+          <Route path='inicio' component={Homepage}></Route>
+          <Route path='informacoes-gerais' component={GeneralInformation}></Route>
+          <Route path='solicitar' component={RequestLocation}></Route>
+          <Route path='receber' component={Receive}></Route>
+          <Route path='historico' component={History}></Route>
+        </Route>
+      </Router>
+    </IntlProvider>
+  </Provider>,
 document.getElementById('root'));
